requirements.txt
appdirs==1.4.4
apturl==0.5.2
astroid==2.9.3
attrs==22.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.11.2
blinker==1.4
Brlapi==0.8.3
ccsm==0.9.14.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
colorlog==6.7.0
command-not-found==0.3
compizconfig-python==0.9.14.1
cppman==0.5.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
flake8==4.0.1
Flask==2.0.1
future==0.18.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isort==5.6.4
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
natsort==8.3.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
online-judge-api-client==10.10.1
online-judge-tools==11.5.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.13
platformdirs==2.5.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.4
systemd-python==234
tk==0.1.0
toml==0.10.2
trash-cli==0.17.1.14
typing-extensions==3.10.0.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
xdg==5
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
